+++ examples/common.styles.js
@@ -1,15 +1,14 @@
-module.exports = {
-  resetListStyle: {
-    listStyle: 'none'
-  },
-  resetBoxModel: {
-    marginTop: 0,
-    marginRight: 0,
-    marginBottom: 0,
-    marginLeft: 0,
-    paddingTop: 0,
-    paddingRight: 0,
-    paddingBottom: 0,
-    paddingLeft: 0
-  }
+export const resetListStyle = {
+  listStyle: 'none'
+};
+
+export const resetBoxModel = {
+  marginTop: 0,
+  marginRight: 0,
+  marginBottom: 0,
+  marginLeft: 0,
+  paddingTop: 0,
+  paddingRight: 0,
+  paddingBottom: 0,
+  paddingLeft: 0
 };