+++ karma.conf.js
@@ -1,5 +1,12 @@
 const path = require('path');
 
+// Note: If we switch to ESM version of babelified files, we'll likely need to
+// update from the ancient isparta-loader. Likely we'll switch to:
+// https://github.com/istanbuljs/babel-plugin-istanbul with a `test` BABEL_ENV
+//
+// https://github.com/FormidableLabs/radium/issues/969
+process.env.BABEL_ENV = 'commonjs';
+
 module.exports = function(config) {
   config.set({
     basePath: '',
@@ -20,7 +27,7 @@ module.exports = function(config) {
     webpack: {
       cache: true,
       module: {
-        loaders: [
+        rules: [
           {
             test: /\.js$/,
             enforce: 'pre',