+++ webpack.config.js
@@ -2,7 +2,7 @@ const path = require('path');
 
 module.exports = {
   cache: true,
-  entry: path.join(__dirname, '/src/index.js'),
+  entry: path.join(__dirname, 'src/index.js'),
   externals: [
     {
       react: {
@@ -14,15 +14,17 @@ module.exports = {
     }
   ],
   output: {
-    path: path.join(__dirname, '/dist'),
+    path: path.join(__dirname, 'dist'),
     filename: 'radium.js',
     library: 'Radium',
-    libraryTarget: 'umd'
+    libraryTarget: 'umd',
+    libraryExport: 'default'
   },
   module: {
-    loaders: [
+    rules: [
       {
         test: /\.js$/,
+        include: [path.join(__dirname, 'src')],
         loader: 'babel-loader'
       }
     ]