+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## UNRELEASED MAJOR (TODO_DATE)
 ### Breaking Changes
-- Radium now exports defaults as `.default`, so for runtimes like Node.js for all files in `lib/**`. We have changed `package.json:main` to point to `/index.js` instead of `/lib/index.js` as a convenience wrapper to expose mostly what was there before so behavior of `const Radium = require('radium');` works mostly as it did before. But, any imports of a default export from a file in lib like `const Enhancer = require("radium/lib/enhancer");` will need to be changed to `const Enhancer = require("radium/lib/enhancer").default;`.
+- Radium now exports defaults as `.default`, so for runtimes like Node.js for all files in `lib/**`. We have changed `package.json:main` to point to `/index.js` instead of `/lib/index.js` as a convenience wrapper to expose mostly what was there before so behavior of `const Radium = require('radium');` works mostly as it did before. But, any imports of a default export from a file in lib like `const Enhancer = require('radium/lib/enhancer');` will need to be changed to `const Enhancer = require('radium/lib/enhancer').default;`.
 
 ### Features
 - Add `es` ESM module export files.